Bug 2027318 - Destroy RemoteSpellcheckEngineChild correctly. r=masayuki
Differential Revision: https://phabricator.services.mozilla.com/D292146

UltraBlame original commit: 158c6bfebb106ae42d5d4e3412e0d41f3865a079

+++ extensions/spellcheck/ipc/RemoteSpellCheckEngineChild.cpp
@@ -13,7 +13,17 @@ RemoteSpellcheckEngineChild::RemoteSpellcheckEngineChild(
 RemoteSpellcheckEngineChild::~RemoteSpellcheckEngineChild() {
   
   
-  mOwner->DeleteRemoteEngine();
+  if (mOwner) {
+    mOwner->DeleteRemoteEngine();
+  }
+}
+
+void RemoteSpellcheckEngineChild::Destroy() {
+  if (mOwner) {
+    mOwner->DeleteRemoteEngine();
+    mOwner = nullptr;
+  }
+  (void)RemoteSpellcheckEngineChild::Send__delete__(this);
 }
 
 RefPtr<GenericPromise> RemoteSpellcheckEngineChild::SetCurrentDictionaries(
